src/main/java/com/company/accountservice/web/AccountsController.java
package com.company.accountservice.web;

import com.company.accountservice.service.AccountService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import static org.springframework.web.bind.annotation.RequestMethod.GET;

@RestController
public class AccountsController {

    private final AccountService accountService;

    @Autowired
    public AccountsController(AccountService accountService) {
        this.accountService = accountService;
    }

    @RequestMapping(path = "/accounts", method = GET)
    public ResponseEntity<?> getAccounts(){
        return ResponseEntity.ok().body(accountService.getAccounts());
    }

}
